requirements.txt
streamlit==1.37.1
PyMuPDF==1.24.9
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
transformers==4.43.3
torch>=2.2.0
reportlab==4.2.2
numpy>=1.26.0
pymongo
passlib[bcrypt]
google-cloud-storage
openai
scikit-learn